Add support for accelerating mean and rms with numpy
The previous implementation used Python's built-in ** operator and
math.sum to calculate the mean and rms values of an AnalogSpan.

On my machine (Ryzen 5950x), this took around 11s to run all
measurements on 30M samples. VoltageMeasurer, which is dominated by the
rms and mean calculations, took 2.18s of the 11s. After replacing with a
numpy implementation, it takes 0.088s.